feat: update connect-ios-swiftui package
