--- connect-ios-swiftui/SampleiOSSwiftUIApp/Assets.xcassets/Color=Winter White.imageset/Contents.json
@@ -1,21 +1,0 @@
-{
-  "images": [
-    {
-      "filename": "Color=Winter White.png",
-      "idiom": "universal",
-      "scale": "1x"
-    },
-    {
-      "idiom": "universal",
-      "scale": "2x"
-    },
-    {
-      "idiom": "universal",
-      "scale": "3x"
-    }
-  ],
-  "info": {
-    "author": "xcode",
-    "version": 1
-  }
-}
