--- connect-ios-swiftui/SampleiOSSwiftUIApp/Info.plist
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>NSAppTransportSecurity</key>
-	<dict>
-		<key>NSExceptionDomains</key>
-		<dict>
-			<key>anon.com</key>
-        <dict>
-            <key>NSIncludesSubdomains</key>
-            <true/>
-            <key>NSExceptionAllowsInsecureHTTPLoads</key>
-            <true/>
-            <key>NSExceptionRequiresForwardSecrecy</key>
-            <true/>
-            <key>NSExceptionMinimumTLSVersion</key>
-            <string>TLSv1.2</string>
-            <key>NSThirdPartyExceptionAllowsInsecureHTTPLoads</key>
-            <false/>
-            <key>NSThirdPartyExceptionRequiresForwardSecrecy</key>
-            <true/>
-            <key>NSThirdPartyExceptionMinimumTLSVersion</key>
-            <string>TLSv1.2</string>
-            <key>NSRequiresCertificateTransparency</key>
-            <false/>
-        </dict>
-		</dict>
-	</dict>
-</dict>
-</plist>
