--- connect-ios-swiftui/SampleiOSSwiftUIApp/Preview Content/Preview Assets.xcassets/Contents.json
@@ -1,6 +1,0 @@
-{
-  "info": {
-    "author": "xcode",
-    "version": 1
-  }
-}
