+++ connect-ios-swiftui/connect-ios-swiftui.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,0 +1,24 @@
+{
+  "originHash" : "bd1531fe8c4ff73dbebca0d1c320d2c9f035d0778de0d23dee44e5dc5315a956",
+  "pins" : [
+    {
+      "identity" : "jwtdecode.swift",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/auth0/JWTDecode.swift.git",
+      "state" : {
+        "revision" : "58af7278797871e460d79496621b3e5366b865b2",
+        "version" : "3.1.0"
+      }
+    },
+    {
+      "identity" : "sdk-ios-pkg",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/anon-dot-com/sdk-ios-pkg",
+      "state" : {
+        "branch" : "main",
+        "revision" : "8eff07260098b15825af65d879bf20019b97cbf2"
+      }
+    }
+  ],
+  "version" : 3
+}
