+++ connect-ios-swiftui/connect-ios-swiftui/connect_ios_swiftuiApp.swift
@@ -1,0 +1,17 @@
+//
+//  connect_ios_swiftuiApp.swift
+//  connect-ios-swiftui
+//
+//  Created by Haoyu Yun on 6/21/24.
+//
+
+import SwiftUI
+
+@main
+struct connect_ios_swiftuiApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
